Unit Of Work Added

+++ src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../_services/confirm.service';
 import { MessageService } from './../_services/message.service';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
@@ -17,7 +18,7 @@ export class MessagesComponent implements OnInit {
   pageSize=5;
   loadingFlag = false;
 
-  constructor(private messageService :MessageService) { }
+  constructor(private messageService :MessageService, private confirmService : ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -35,9 +36,15 @@ export class MessagesComponent implements OnInit {
 
   deleteMessage(id : number)
   {
-    this.messageService.deleteMessage(id).subscribe(() =>{
-      this.messages.splice(this.messages.findIndex(m=> m.id === id),1);
+    this.confirmService.confirm("Confirm delete message", "This cannot be undone").subscribe(result => {
+      if(result)
+      {
+        this.messageService.deleteMessage(id).subscribe(() =>{
+          this.messages.splice(this.messages.findIndex(m=> m.id === id),1);
+        });
+      }
     })
+  
 
   }
